requirements.txt
django
djangorestframework
snowflake-sqlalchemy
pandas